Introduce core classes plugin; remove corresponding code from Layout
